+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.0
+
+* renamed encoding and decoding option `check_symbol` to `checksum`.
+* removed `partition_length` encoding option.
+* introduced `partitions` encoding option.
+* improved documentation.
+
 ## v0.1.0
 
 * basic functionality